requirements.txt
langchain==0.1.0
langchain-community==0.0.20
langchain-perplexity==1.0.0
streamlit==1.29.0
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.0
sentence-transformers==2.2.2
